ST2->ST3: Convert WebosCommand.view_output() to WebosViewOutputCommand
In Sublime Text 3 it's no longer possible to start your own edit, which
`view_output()` relied on, so it has instead been made into
WebosViewOutputCommand which can achieve the same thing since it's based on the
`sublime_plugin.TextCommand` class.

There's also a few minor fixes to wording and such.