Deduplicate code for creating applications and services
There's a limited difference between the workflows of creating apps
and services, so we might as well just make them the same function
with a single `if` statement to differentiate.

It will also make it easier to add support for other types of project
in the future, as it's just a matter of supplying a variable to the
command.

This commit also adds more information about each template to the
selection list and adds the option of where the project should be
created, plus somewhat sensitive defaults.